+++ CHANGES.rst
@@ -67,6 +67,7 @@ Changelog
   Definable TAL expressions make adaptable method MeetingItem._isViewableByPowerObservers obsolete, it is removed.
 - Removed unused fields on ToolPloneMeeting : 'extractTextFromFiles', 'availableOcrLanguages', 'defaultOcrLanguage' and 'enableUserPreferences'
 - Added possibility to define a Plone group in the recipient on a mailing_list defined on a POD template using 'group:' + Plone group id
+- Make sure a MeetingCategory can not be deleted if used in field 'categoryMappingsWhenCloningToOtherMC' of another MeetingCategory in another MeetingConfig
 
 4.1b16 (2019-01-31)
 -------------------